requirements.txt
# Core dependencies
python-dotenv==1.1.1
streamlit==1.47.1
openai==1.97.1

# Data analysis and visualization
pandas==2.3.1
matplotlib==3.10.3
numpy==1.26.4
seaborn==0.13.2
scipy==1.11.4
scikit-learn==1.3.2
statsmodels==0.14.1

# PDF report generation
reportlab==4.0.7
weasyprint==60.2
jinja2==3.1.2

# Optional: LangChain components (if using the full agent implementation)
langchain==0.1.9
langchain-openai==0.1.5
langgraph==0.0.46
pydantic==1.10.13

# SQL database connection
# Note: Uncomment and specify the appropriate database driver based on your needs
# psycopg2-binary==2.9.9  # For PostgreSQL
# pymysql==1.1.0         # For MySQL
# sqlite3                # Built into Python (no need to install)
